requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI
openai>=1.30.0

# Korean NLP
konlpy==0.6.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CORS
starlette==0.35.1

